requirements.txt
python-docx==0.8.6
requests==2.17.3
